requirements.txt
langchain>=0.1.0
langchain-core>=0.1.4
langgraph>=0.0.19
tavily-python>=0.2.6
python-dotenv>=1.0.0
google-generativeai>=0.3.2
beautifulsoup4>=4.12.2
requests>=2.31.0
pytest>=7.4.0